requirements.txt

pandas==2.1.0
numpy==1.24.0
requests==2.31.0
tqdm==4.66.0
sentence-transformers==2.6.0
umap-learn==0.5.6
hdbscan==0.8.33
anthropic==0.25.0
fastapi==0.110.0
uvicorn==0.29.0
aiofiles==23.2.1
scikit-learn==1.4.0
